+++ src/rollup/config.ts
@@ -1,7 +1,7 @@
 import { pathToFileURL } from 'url'
 import { dirname, join, normalize, relative, resolve } from 'pathe'
 import type { InputOptions, OutputOptions } from 'rollup'
-import defu from 'defu'
+import { defu } from 'defu'
 import { terser } from 'rollup-plugin-terser'
 import commonjs from '@rollup/plugin-commonjs'
 import { nodeResolve } from '@rollup/plugin-node-resolve'
@@ -57,7 +57,7 @@ export const getRollupConfig = (nitro: Nitro) => {
   const buildServerDir = join(nitro.options.buildDir, 'dist/server')
   const runtimeAppDir = join(runtimeDir, 'app')
 
-  const rollupConfig: RollupConfig = defu<RollupConfig, RollupConfig>(nitro.options.rollupConfig, {
+  const rollupConfig = defu(nitro.options.rollupConfig, <RollupConfig> {
     input: nitro.options.entry,
     output: {
       dir: nitro.options.output.serverDir,
@@ -121,8 +121,8 @@ export const getRollupConfig = (nitro: Nitro) => {
     rollupConfig.plugins.push(timing())
   }
 
-  if (nitro.options.autoImport) {
-    rollupConfig.plugins.push(unimportPlugin.rollup(nitro.options.autoImport))
+  if (nitro.options.imports) {
+    rollupConfig.plugins.push(unimportPlugin.rollup(nitro.options.imports))
   }
 
   // Raw asset loader
@@ -255,6 +255,7 @@ export const plugins = [
       },
       exportConditions: [
         'default',
+        nitro.options.dev ? 'development' : 'production',
         'module',
         'node',
         'import'
@@ -282,6 +283,7 @@ export const plugins = [
     mainFields: ['main'],
     exportConditions: [
       'default',
+      nitro.options.dev ? 'development' : 'production',
       'module',
       'node',
       'import'