+++ src/utils/index.ts
@@ -67,6 +67,7 @@ export function replaceAll (input: string, from: string, to: string) {
 
 const autodetectableProviders = {
   azure_static: 'azure',
+  cloudflare_pages: 'cloudflare_pages',
   netlify: 'netlify',
   stormkit: 'stormkit',
   vercel: 'vercel'
@@ -132,7 +133,12 @@ export function readPackageJson (
   }
 }
 
-export function resolveAliases (aliases: Record<string, string>) {
+export function resolveAliases (_aliases: Record<string, string>) {
+  // Sort aliases from specific to general (ie. fs/promises before fs)
+  const aliases = Object.fromEntries(Object.entries(_aliases).sort(([a], [b]) =>
+    (b.split('/').length - a.split('/').length) || (b.length - a.length)
+  ))
+  // Resolve alias values in relation to each other
   for (const key in aliases) {
     for (const alias in aliases) {
       if (!['~', '@', '#'].includes(alias[0])) { continue }