+++ test/fixture/nitro.config.ts
@@ -1,7 +1,7 @@
 import { defineNitroConfig } from '../../src'
 
 export default defineNitroConfig({
-  autoImport: {
+  imports: {
     presets: [
       {
         // TODO: move this to built-in preset
@@ -22,6 +22,9 @@ export default defineNitroConfig({
   ],
   prerender: {
     crawlLinks: true,
+    ignore: [
+      // '/api/param/'
+    ],
     routes: [
       '/prerender',
       '/404'